model/requirements.txt
fastapi
uvicorn
torch
transformers
huggingface_hub
Pillow
python-multipart